Functionality to easy and hard buttons
Made it so if you click the easy button and it is currently in hard mode, then it will start a new easy game. IF it's already in easy mode then it won't do anything (if you wanted new easy game hit new colors). And the same functionality is done to the hard button too. One thing I have not done is hide the last 3 colors if in easy mode.