Fnished the bulk of the style
FInished most of the 'core' styling. Big things that remaind to do are have the colors dissapear if they are clicked and aren't the winning color. Need to change the background/font color change of whatever game mode is toggled. Want a hover effect on new collors button. NEed to update message appropriately, it is static right now. And, I want to research into responsive desgin/have this compatible on mobile as a final touch.